jdvrlib: playback fast on starting [1/1]
PD#SWPL-149293
PD#SWPL-144801

Problem:
PVR playback is in fast speed shortly on starting.
Injecting video data twice to video decoder would cause the issue.
The purpose of first inject is for PMT acquisition. Only PSI data
is expected to be injected. Video data is not required to be injected
to video decoder on this situation.

Solution:
A hack to make the AV data null packets at first injection
so that they will not be injected to video decoder.

Verify:
Tested pass on ohm

Signed-off-by: Wentao.MA <wentao.ma@amlogic.com>
Change-Id: I3acded39da871a21a94ddce4e4475f7061012b7a