jdvrlib: playback stuck in seek [1/1]
PD#SWPL-152734
PD#SWPL-155624

Problem:
JDvrPlayer fails to determine playing time based on received PTS
occasionally in seeking. It is known that PTS in streams are not
linear inceasing. When playing time was updated with an earlier
and higher PTS value, JDvrPlayer is not able to calculate playing
time based on a later but lower PTS value. The index finding
algorithm need to be adjusted to accommodate this situation.

Solution:
Move index finding reference point 1000ms backward to increase the
hit rate.

Verify:
Tested pass on ohm

Signed-off-by: Wentao.MA <wentao.ma@amlogic.com>
Change-Id: I47f56188e90e4060bc275c2c470b00fa15688758