jdvrlib: rewind to starting position and it will not play automatically [1/1]
PD#SWPL-160692

Problem:
Rewind playback at 48x speed until it reaches beginning point.
Sometimes, it stuck here and cannot resume to normal speed.
Reason 1, like SWPL-149627, ASPlayer may fail to notify the
PTS within the first index of the recording, so JDvrPlayer fails
to notify the 0 playing time to higher layer for it to resume
normal speed.
Reason 2, a defect in index searching algorithm.

Solution:
1. Introduce 1000ms EXIT_THRESHOLD like SWPL-149627's fix and set
playing time to 0 if the playback is in rewind and has reached
beginning.
2. Fix the defect in index searching algorithm.

Verify:
Tested pass on ohm

Change-Id: I610cc76036ae4d405025929ab70797b7ce426b4c
Signed-off-by: Wentao.MA <wentao.ma@amlogic.com>