jdvrlib: FF to the end will not automatically exit [1/1]
PD#SWPL-161192

Problem:
Fast forward to the end but playback will not exit automatically.
The reason is that last skip doesn't inject enough data to ASPlayer
to generate pts at endpoint.

Solution:
Last skip to the 1000ms from the endpoint which is expected to
generate pts within the last 1000ms for JDvrPlayer to know that
playback reaches end.

Verify:
Tested pass on ohm

Change-Id: I3a1c8dc70e398e854cad154b017092bff6074094
Signed-off-by: Wentao.MA <wentao.ma@amlogic.com>